🗑️ [Del]+🐛[Fix]: 파일 삭제 + 이름 변경
	* CircularSlider.swift 삭제
	* pre)CircularSlider -> CircularSlider로 변경
