+++ RelaxOn/App/Views/Components/CircularSlider.swift
@@ -15,29 +15,34 @@ import SwiftUI
 struct CircularSlider: View {
     
     @EnvironmentObject var viewModel: CustomSoundViewModel
+    // 회전각도 관련 속성
     @State var angle: Double = Double.random(in: 0...360)
+    // 이미지의 위치와 방향을 정하는 속성
+    @State private var rotationAngle = Angle(degrees: 0)
+    var angleChanged: (Double) -> Void
+    
     @State var type: CircleType
     @State private var currentFilterIndex = 0
     @State private var filter: AudioFilter
     @State private var filters: [AudioFilter] = []
-    private var width: CGFloat {
-        type.width
-    }
     
     var imageName: String
-    
     var isOnMove: Bool = true
-    var angleChanged: (Double) -> Void
-    var range: [Float]
+    private var minValue = 0.0
+    private var maxValue = 1.0
+    private var width: CGFloat { type.width }
+    
+    
     
     // 슬라이더의 angle값을 반환
     init(type: CircleType, imageName: String, gestureType: Bool, range: [Float], filter: AudioFilter = .WaterDrop, angleChanged: @escaping (Double) -> Void) {
         self.type = type
         self.imageName = imageName
         self.isOnMove = gestureType
-        self.range = range
-        self._filter = State(initialValue: filter)
         self.angleChanged = angleChanged
+        self._filter = State(initialValue: filter)
+        self.minValue = Double(range.first ?? 0)
+        self.maxValue = Double(range.last ?? 1)
     }
     
     var body: some View {
@@ -46,21 +51,24 @@ struct CircularSlider: View {
             .resizable()
             .scaledToFit()
             .frame(width: 24)
-            .foregroundColor(Color.systemGrey3)
-            .rotationEffect(.init(degrees: -angle))
+        // FIXME: rotationAngle이 +90이 되는 현상
+            .rotationEffect(-rotationAngle + Angle(degrees: 90))
             .offset(x: width / 2)
-            .rotationEffect(.init(degrees: angle))
+            .rotationEffect(rotationAngle - Angle(degrees: 90))
             .gesture(
-                DragGesture()
-                    .onChanged({ value in
+                DragGesture(minimumDistance: 0.0)
+                    .onChanged(){ value in
                         if isOnMove {
-                            onDrag(value: value)
+                            onDrag(value: value.location)
+                            print("Angle : \(angle)")
+                            print("Rotation : \(rotationAngle)")
                         } else {
-                            onMove(value: value)
+                            onMove(value: value.location)
                         }
-                    })
+                    }
             )
             .onAppear {
+                self.rotationAngle = Angle(degrees: progressFraction * 360.0)
                 if let filterArray = viewModel.filterDictionary[filter] {
                     filters.append(contentsOf: filterArray)
                 }
@@ -76,28 +84,53 @@ struct CircularSlider: View {
      SnappedAngle은 정해진 각도로만 이동하기위한 함수
      */
     // 슬라이드형 움직임
-    func onDrag(value: DragGesture.Value) {
-        let vector = CGVector(dx: value.location.x, dy: value.location.y)
-        let radians = atan2(vector.dy, vector.dx)
-        let angle = radians * 180 / .pi
-        self.angle = angle
+    func onDrag(value: CGPoint) {
+        // 입력 받은 위치로 벡터를 생성합니다. (iOS는 y축이 반대 방향이므로 -y로 설정합니다.)
+        let vector = CGVector(dx: value.x, dy: -value.y)
+        
+        // atan2 함수를 사용하여 벡터의 각도를 계산합니다.
+        let angleRadians = atan2(vector.dx, vector.dy)
+        
+        // 각도가 음수인 경우를 대비해, 각도를 0 ~ 2π 범위로 맞춥니다.
+        let positiveAngle = angleRadians < 0.0 ? angleRadians + (2.0 * .pi) : angleRadians
+        
+        // 계산된 각도를 이용해서 angle 값을 업데이트합니다.
+        angle = ((positiveAngle /  (2.0 * .pi)) * (maxValue - minValue )) + minValue
         angleChanged(angle)
+        
+        rotationAngle = Angle(radians: positiveAngle)
     }
     
     // 이동형 움직임
-    func onMove(value: DragGesture.Value) {
-        let vector = CGVector(dx: value.location.x, dy: value.location.y)
-        let radians = atan2(vector.dy - width / 2, vector.dx - width / 2)
-        let angle = radians * 180 / .pi
+    func onMove(value: CGPoint) {
+        // 입력 받은 위치로 벡터를 생성합니다. (iOS는 y축이 반대 방향이므로 -y로 설정합니다.)
+        let vector = CGVector(dx: value.x, dy: -value.y)
+        
+        // atan2 함수를 사용하여 벡터의 각도를 계산합니다.
+        let angleRadians = atan2(vector.dx, vector.dy)
+        
+        // 각도가 음수인 경우를 대비해, 각도를 0 ~ 2π 범위로 맞춥니다.
+        let positiveAngle = angleRadians < 0.0 ? angleRadians + (2.0 * .pi) : angleRadians
+        
+        // 계산된 각도를 이용해서 angle 값을 업데이트합니다.
+        angle = ((positiveAngle /  (2.0 * .pi)) * (maxValue - minValue )) + minValue
+        angleChanged(angle)
+        rotationAngle = Angle(radians: positiveAngle)
+        
+        
         let snappedAngle = round(angle / 72) * 72
-
         if snappedAngle != self.angle {
             currentFilterIndex = (currentFilterIndex + 1) % filters.count
         }
         
         self.angle = Double(snappedAngle)
-        angleChanged(angle)
         updateFilter()
+    }
+    
+    // 진행률을 계산하는 private 변수입니다.
+    private var progressFraction: Double {
+        // 진행률은 현재 값에서 최소값을 빼고, 그 결과를 (최대값 - 최소값)으로 나눈 값입니다.
+        return ((angle - minValue) / (maxValue - minValue))
     }
     
     func updateFilter() {
@@ -112,7 +145,7 @@ struct CircularSlider: View {
     }
 }
 
-struct CircularSlider_Previews: PreviewProvider {
+struct preCircularSliderView_Previews: PreviewProvider {
     static var previews: some View {
         CircularSlider(type: .medium, imageName: "filter", gestureType: true, range: [1.0]) { _ in }
             .environmentObject(CustomSoundViewModel())
